+++ lib/grape/util/content_types.rb
@@ -5,10 +5,8 @@ module Grape
       :xml,  'application/xml',
       :serializable_hash, 'application/json',
       :json, 'application/json',
-      :jsonapi, 'application/vnd.api+json',
-      :atom, 'application/atom+xml',
-      :rss,  'application/rss+xml',
-      :txt,  'text/plain',
+      :binary, 'application/octet-stream',
+      :txt,  'text/plain'
    ]
 
     def self.content_types_for(from_settings)